requirements.txt
gunicorn==19.3.0
alabaster==0.7.6
Babel==2.0
blinker==1.4
click==5.1
docutils==0.12
dominate==2.1.12
Flask==0.10.1
Flask-Babel==0.9
Flask-Bootstrap==3.3.5.6
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Moment==0.5.1
flask-mongoengine==0.7.1
flask-paginate==0.4.5
Flask-PyMongo==0.3.1
Flask-Script==2.0.5
Flask-Session==0.2.1
Flask-WTF==0.12
Foundations==2.1.0
itsdangerous==0.24
Jinja2==2.8
linecache2==1.0.0
Manager==2.0.5
MarkupSafe==0.23
mongoengine==0.10.0
Oncilla==0.1.0
Pygments==2.0.2
pymongo==3.0.3
python-dotenv==0.1.3
pytz==2015.4
six==1.9.0
snowballstemmer==1.2.0
speaklater==1.3
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
traceback2==1.4.0
unittest2==1.1.0
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
